requirements.txt
requests>=2.31.0
selenium>=4.15.0
Pillow>=10.0.0
rich>=13.7.0
humanize>=4.9.0
pyinstaller>=6.0.0